VLOG-16 [IMP] Rename to Vulcan Utils
