+++ docs/README.md
@@ -1,9 +1,9 @@
-# Vulcan Logger
-Vulcan Logger is a Python package that provides a customizable logging utility with support for automatic inclusion of caller's filename and line number in logs. It aims to simplify logging in Python applications by offering features such as log level configuration, colored logs, and conditional logging.
+# Vulcan Utils
+Vulcan Utils is a Python package that provides a customizable logging utility with support for automatic inclusion of caller's filename and line number in logs. It aims to simplify logging in Python applications by offering features such as log level configuration, colored logs, and conditional logging.
 
 [View the full documentation here](https://vulcan-logger.readthedocs.io/en/latest/)
 
-![Vulcan Logger](https://raw.githubusercontent.com/nodadyoushutup/vulcan-logger/main/docs/img/examples.png)
+![Vulcan Utils](https://raw.githubusercontent.com/nodadyoushutup/vulcan-logger/main/docs/img/examples.png)
 
 ## Features
 - **Caller Information**: Automatically includes the caller's filename and line number in log messages for better traceability.
@@ -12,7 +12,7 @@ Vulcan Logger is a Python package that provides a customizable logging utility w
 - **Conditional Logging**: Conditionally log messages based on specific conditions to control verbosity.
 
 ## Installation
-You can install Vulcan Logger via PIP:
+You can install Vulcan Utils via PIP:
 
 ```bash
 pip install vulcan-logger
@@ -22,12 +22,12 @@ pip install vulcan-logger
 [View example usage here](https://github.com/nodadyoushutup/vulcan-logger/blob/main/example/example.py)
 
 ### Basic Logging
-1. Import the Logger class from the `vulcan_logger.logger` module.
+1. Import the Logger class from the `vulcan_utils.logger` module.
 2. Initialize a Logger instance with a name and an optional log level. The default log level is INFO.
 3. Use the logging methods (debug, info, warning, error, critical) to log messages at various severity levels.
 
 ```python
-from vulcan_logger.logger import Logger
+from vulcan_utils.logger import Logger
 
 # Initialize the logger with a custom name and log level
 logger = Logger(name='example', level='DEBUG')
@@ -41,13 +41,13 @@ logger.critical("Critical message for severe conditions")
 ```
 
 ### Function Logging with Decorators
-Vulcan Logger provides a logging decorator that can be applied to functions to automatically log calls, returns, and execution times. To use this feature:
+Vulcan Utils provides a logging decorator that can be applied to functions to automatically log calls, returns, and execution times. To use this feature:
 
-1. Import the log decorator from the `vulcan_logger.decorator` module.
+1. Import the log decorator from the `vulcan_utils.decorator` module.
 2. Apply the @log decorator to any function. You can optionally specify a log level or a condition for logging.
 
 ```python
-from vulcan_logger.decorator import log
+from vulcan_utils.decorator import log
 
 @log(level="DEBUG")
 def compute_sum(a, b):
@@ -65,7 +65,7 @@ product_result = conditional_log_example(5, 3)
 ```
 
 ### Advanced Configuration
-Vulcan Logger offers several environment variables to fine-tune logging behavior for your application. You can set these variables before initializing your logger to customize logging output, destination, and file naming.
+Vulcan Utils offers several environment variables to fine-tune logging behavior for your application. You can set these variables before initializing your logger to customize logging output, destination, and file naming.
 
 #### Setting Global Log Level
 Control the log level globally across your application by setting the `VULCAN_LOG_LEVEL` environment variable. This determines the minimum level of messages that will be logged. Available levels are `DEBUG`, `INFO`, `WARNING`, `ERROR`, and `CRITICAL`.
@@ -81,7 +81,7 @@ os.environ["VULCAN_LOG_LEVEL"] = "WARNING"
 ```
 
 #### Setting Log File Path
-By default, Vulcan Logger writes logs to the current directory. Set the `VULCAN_LOG_PATH` environment variable to specify a custom directory for log files.
+By default, Vulcan Utils writes logs to the current directory. Set the `VULCAN_LOG_PATH` environment variable to specify a custom directory for log files.
 
 _bash_
 ```bash
@@ -107,7 +107,7 @@ os.environ["VULCAN_LOG_NAME"] = "example"
 ```
 
 ### Handling Exceptions
-Vulcan Logger makes it easy to log exceptions. Use the logging methods within exception handling blocks to log errors and critical issues.
+Vulcan Utils makes it easy to log exceptions. Use the logging methods within exception handling blocks to log errors and critical issues.
 
 ```python
 try:
@@ -118,7 +118,7 @@ except ZeroDivisionError as e:
 ```
 
 ## Contributing
-Contributions to Vulcan Logger are welcome! To contribute, follow these steps:
+Contributions to Vulcan Utils are welcome! To contribute, follow these steps:
 
 1. Fork the repository and clone it to your local machine.
 2. Install the development dependencies by running `pip install -r requirements.txt`.
