+++ docs/source/index.rst
@@ -1,8 +1,8 @@
 .. docs/source/index.rst
-.. Vulcan Logger documentation master file, created by sphinx-quickstart on Fri Apr 5 22:45:24 2024.
+.. Vulcan Utils documentation master file, created by sphinx-quickstart on Fri Apr 5 22:45:24 2024.
    You can adapt this file completely to your liking, but it should at least contain the root `toctree` directive.
 
-Welcome to Vulcan Logger's documentation!
+Welcome to Vulcan Utils's documentation!
 =========================================
 
 .. toctree::
